project/requirements.txt
astroid==2.3.2
dj-database-url==0.5.0
Django==2.2.6
django-crispy-forms==1.8.0
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==20.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.3
pandas==0.25.3
Pillow==6.2.0
pkg-resources==0.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
Pygments==2.4.2
pylint==2.4.3
python-dateutil==2.8.1
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.4
wrapt==1.11.2
